added axios, changeв register form and form group
