--- src/client/Components/FormGroup/FormGroupValidate.js
@@ -1,43 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const FormGroupValidate = ({
-  htmlFor,
-  label,
-  type,
-  className,
-  id,
-  placeholder,
-  name,
-  value,
-  onChange,
-}) => (
-  <div className="form-group">
-    <label htmlFor={htmlFor}>
-      {label}
-      <input
-        type={type}
-        className={className}
-        id={id}
-        placeholder={placeholder}
-        name={name}
-        value={value}
-        onChange={onChange}
-      />
-    </label>
-  </div>
-);
-
-FormGroupValidate.propTypes = {
-  htmlFor: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-  type: PropTypes.string.isRequired,
-  className: PropTypes.string.isRequired,
-  id: PropTypes.string.isRequired,
-  placeholder: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  value: PropTypes.string.isRequired,
-  onChange: PropTypes.func.isRequired,
-};
-
-export default FormGroupValidate;
